refactor(handlers): extract admin handlers (WP5.9.5.1)
Extract admin-related handlers into separate useAdminHandlers hook:
- handleClearAllCourts
- handleAdminClearCourt
- handleMoveCourt
- handleClearWaitlist
- handleRemoveFromWaitlist
- handlePriceUpdate
- handleExitAdmin

useRegistrationHandlers.js: 1,213 → 1,156 lines (-57 lines)
adminHandlers.js: 110 lines (new)
