+++ src/registration/appHandlers/handlers/index.js
@@ -1,0 +1,1 @@
+export { useAdminHandlers } from './adminHandlers.js';
